Dijkstra algorithm implemented. Final Scene Updated according to Tutorial.
I was working on a Tutorial and use the variable that was in the
tutorial. Code has still needs to be cleaned and real time walkable
tiles needs possible to be changed. New materials and Prefabs are made
due to the Tutorial. I will optimize the usage without prefabs
implementation through inspector.