Commit with new menubackground ect.
This commit has the new new background for the menu, the new locked
level is also fixed, sounds for the enemies are also fixed BUT STILL
COMMENTED, music fixed for the next waves and level. New sliders were
also added to option menu, fonts were also added as well as the options
background for Options of course.